refactor(pm): split clap definitions into cli.rs, make main.rs a thin dispatcher
main.rs had absorbed the dispatcher role plus per-command business
assembly (812 lines). The clap derives and parsing tests move to a new
cli.rs; the Install/Uninstall assembly (omit folding, global-vs-local
branching, save-type, pluralized logging), the --from pnpm pre-step,
the None-arm config-command/script fallback, and the config/link/deps
dispatch all move into their owning cmd/ modules as run() entries, so
every match arm is a 1-3 line delegate and main.rs is 250 lines of
bootstrap + match. The cmd-layer duplicate of the service's
specs-empty guard is gone (the service check is authoritative).
